Swift 2.0 and Xcode 7beta update!
