+++ Design-Patterns.playground/Pages/Behavioral.xcplaygroundpage/Contents.swift
@@ -1,0 +1,608 @@
+//: Behavioral |
+//: [Creational](Creational) |
+//: [Structural](Structural)
+/*:
+Behavioral
+==========
+
+>In software engineering, behavioral design patterns are design patterns that identify common communication patterns between objects and realize these patterns. By doing so, these patterns increase flexibility in carrying out this communication.
+>
+>**Source:** [wikipedia.org](http://en.wikipedia.org/wiki/Behavioral_pattern)
+*/
+import Swift
+import Foundation/*:
+🐝 Chain Of Responsibility
+--------------------------
+
+The chain of responsibility pattern is used to process varied requests, each of which may be dealt with by a different handler.
+
+### Example:
+*/
+class MoneyPile {
+    let value: Int
+    var quantity: Int
+    var nextPile: MoneyPile?
+
+    init(value: Int, quantity: Int, nextPile: MoneyPile?) {
+        self.value = value
+        self.quantity = quantity
+        self.nextPile = nextPile
+    }
+
+    func canWithdraw(var v: Int) -> Bool {
+
+        func canTakeSomeBill(want: Int) -> Bool {
+            return (want / self.value) > 0
+        }
+
+        var q = self.quantity
+
+        while canTakeSomeBill(v) {
+
+            if q == 0 {
+                break
+            }
+
+            v -= self.value
+            q -= 1
+        }
+
+        if v == 0 {
+            return true
+        } else if let next = self.nextPile {
+            return next.canWithdraw(v)
+        }
+
+        return false
+    }
+}
+
+class ATM {
+    private var hundred: MoneyPile
+    private var fifty: MoneyPile
+    private var twenty: MoneyPile
+    private var ten: MoneyPile
+
+    private var startPile: MoneyPile {
+        return self.hundred
+    }
+
+    init(hundred: MoneyPile,
+           fifty: MoneyPile,
+          twenty: MoneyPile,
+             ten: MoneyPile) {
+
+        self.hundred = hundred
+        self.fifty = fifty
+        self.twenty = twenty
+        self.ten = ten
+    }
+
+    func canWithdraw(value: Int) -> String {
+        return "Can withdraw: \(self.startPile.canWithdraw(value))"
+    }
+}
+/*:
+### Usage
+*/
+// Create piles of money and link them together 10 < 20 < 50 < 100.**
+let ten = MoneyPile(value: 10, quantity: 6, nextPile: nil)
+let twenty = MoneyPile(value: 20, quantity: 2, nextPile: ten)
+let fifty = MoneyPile(value: 50, quantity: 2, nextPile: twenty)
+let hundred = MoneyPile(value: 100, quantity: 1, nextPile: fifty)
+
+// Build ATM.
+var atm = ATM(hundred: hundred, fifty: fifty, twenty: twenty, ten: ten)
+atm.canWithdraw(310) // Cannot because ATM has only 300
+atm.canWithdraw(100) // Can withdraw - 1x100
+atm.canWithdraw(165) // Cannot withdraw because ATM doesn't has bill with value of 5
+atm.canWithdraw(30)  // Can withdraw - 1x20, 2x10
+/*:
+👫 Command
+----------
+
+The command pattern is used to express a request, including the call to be made and all of its required parameters, in a command object. The command may then be executed immediately or held for later use.
+
+### Example:
+*/
+protocol DoorCommand {
+    func execute() -> String
+}
+
+class OpenCommand : DoorCommand {
+    let doors:String
+
+    required init(doors: String) {
+        self.doors = doors
+    }
+
+    func execute() -> String {
+        return "Opened \(doors)"
+    }
+}
+
+class CloseCommand : DoorCommand {
+    let doors:String
+
+    required init(doors: String) {
+        self.doors = doors
+    }
+
+    func execute() -> String {
+        return "Closed \(doors)"
+    }
+}
+
+class HAL9000DoorsOperations {
+    let openCommand: DoorCommand
+    let closeCommand: DoorCommand
+
+    init(doors: String) {
+        self.openCommand = OpenCommand(doors:doors)
+        self.closeCommand = CloseCommand(doors:doors)
+    }
+
+    func close() -> String {
+        return closeCommand.execute()
+    }
+
+    func open() -> String {
+        return openCommand.execute()
+    }
+}
+/*:
+### Usage:
+*/
+let podBayDoors = "Pod Bay Doors"
+let doorModule = HAL9000DoorsOperations(doors:podBayDoors)
+
+doorModule.open()
+doorModule.close()
+/*:
+🎶 Interpreter
+--------------
+
+The interpreter pattern is used to evaluate sentences in a language.
+
+### Example
+*/
+
+protocol IntegerExp {
+    func evaluate(context: IntegerContext) -> Int
+    func replace(character: Character, integerExp: IntegerExp) -> IntegerExp
+    func copy() -> IntegerExp
+}
+
+class IntegerContext {
+    private var data: [Character:Int] = [:]
+
+    func lookup(name: Character) -> Int {
+        return self.data[name]!
+    }
+
+    func assign(integerVarExp: IntegerVarExp, value: Int) {
+        self.data[integerVarExp.name] = value
+    }
+}
+
+class IntegerVarExp: IntegerExp {
+    let name: Character
+
+    init(name: Character) {
+        self.name = name
+    }
+
+    func evaluate(context: IntegerContext) -> Int {
+        return context.lookup(self.name)
+    }
+
+    func replace(name: Character, integerExp: IntegerExp) -> IntegerExp {
+        if name == self.name {
+            return integerExp.copy()
+        } else {
+            return IntegerVarExp(name: self.name)
+        }
+    }
+
+    func copy() -> IntegerExp {
+        return IntegerVarExp(name: self.name)
+    }
+}
+
+class AddExp: IntegerExp {
+    private var operand1: IntegerExp
+    private var operand2: IntegerExp
+
+    init(op1: IntegerExp, op2: IntegerExp) {
+        self.operand1 = op1
+        self.operand2 = op2
+    }
+
+    func evaluate(context: IntegerContext) -> Int {
+        return self.operand1.evaluate(context) + self.operand2.evaluate(context)
+    }
+
+    func replace(character: Character, integerExp: IntegerExp) -> IntegerExp {
+        return AddExp(op1: operand1.replace(character, integerExp: integerExp),
+            op2: operand2.replace(character, integerExp: integerExp))
+    }
+
+    func copy() -> IntegerExp {
+        return AddExp(op1: self.operand1, op2: self.operand2)
+    }
+}
+/*:
+### Usage
+*/
+var expression: IntegerExp?
+var intContext = IntegerContext()
+
+var a = IntegerVarExp(name: "A")
+var b = IntegerVarExp(name: "B")
+var c = IntegerVarExp(name: "C")
+
+expression = AddExp(op1: a, op2: AddExp(op1: b, op2: c)) // a + (b + c)
+
+intContext.assign(a, value: 2)
+intContext.assign(b, value: 1)
+intContext.assign(c, value: 3)
+
+var result = expression?.evaluate(intContext)
+/*:
+🍫 Iterator
+-----------
+
+The iterator pattern is used to provide a standard interface for traversing a collection of items in an aggregate object without the need to understand its underlying structure.
+
+### Example:
+*/
+struct NovellasCollection<T> {
+    let novellas: [T]
+}
+
+extension NovellasCollection: SequenceType {
+    typealias Generator = AnyGenerator<T>
+
+    func generate() -> AnyGenerator<T> {
+        var i = 0
+        return anyGenerator{ return i >= self.novellas.count ? nil : self.novellas[i++] }
+    }
+}
+/*:
+### Usage
+*/
+let greatNovellas = NovellasCollection(novellas:["Mist"])
+
+for novella in greatNovellas {
+    print("I've read: \(novella)")
+}
+/*:
+💐 Mediator
+-----------
+
+The mediator pattern is used to reduce coupling between classes that communicate with each other. Instead of classes communicating directly, and thus requiring knowledge of their implementation, the classes send messages via a mediator object.
+
+### Example
+*/
+
+class Colleague {
+    let mediator: Mediator
+
+    init(mediator: Mediator) {
+        self.mediator = mediator
+    }
+
+    func send(message: String) {
+        mediator.send(message, colleague: self)
+    }
+
+    func receive(message: String) {
+        assert(false, "Method should be overriden")
+    }
+}
+
+protocol Mediator {
+    func send(message: String, colleague: Colleague)
+}
+
+class MessageMediator: Mediator {
+    private var colleagues: [Colleague] = []
+
+    func addColleague(colleague: Colleague) {
+        colleagues.append(colleague)
+    }
+
+    func send(message: String, colleague: Colleague) {
+        for c in colleagues {
+            if c !== colleague { //for simplicity we compare object references
+                colleague.receive(message)
+            }
+        }
+    }
+}
+
+class ConcreteColleague: Colleague {
+    override func receive(message: String) {
+        print("Colleague received: \(message)")
+    }
+}
+
+/*:
+### Usage
+*/
+
+let messagesMediator = MessageMediator()
+let user0 = ConcreteColleague(mediator: messagesMediator)
+let user1 = ConcreteColleague(mediator: messagesMediator)
+messagesMediator.addColleague(user0)
+messagesMediator.addColleague(user1)
+
+user0.send("Hello") // user1 receives message
+/*:
+💾 Memento
+----------
+
+The memento pattern is used to capture the current state of an object and store it in such a manner that it can be restored at a later time without breaking the rules of encapsulation.
+
+### Example
+*/
+typealias Memento = Dictionary<NSObject, AnyObject>
+
+let DPMementoKeyChapter = "com.valve.halflife.chapter"
+let DPMementoKeyWeapon = "com.valve.halflife.weapon"
+let DPMementoGameState = "com.valve.halflife.state"
+/*:
+Originator
+*/
+class GameState {
+    var chapter: String = ""
+    var weapon: String = ""
+
+    func toMemento() -> Memento {
+        return [ DPMementoKeyChapter:chapter, DPMementoKeyWeapon:weapon ]
+    }
+
+    func restoreFromMemento(memento: Memento) {
+        chapter = memento[DPMementoKeyChapter] as? String ?? "n/a"
+        weapon = memento[DPMementoKeyWeapon] as? String ?? "n/a"
+    }
+}
+/*:
+Caretaker
+*/
+class CheckPoint {
+    class func saveState(memento: Memento, keyName: String = DPMementoGameState) {
+        let defaults = NSUserDefaults.standardUserDefaults()
+        defaults.setObject(memento, forKey: keyName)
+        defaults.synchronize()
+    }
+
+    class func restorePreviousState(keyName keyName: String = DPMementoGameState) -> Memento {
+        let defaults = NSUserDefaults.standardUserDefaults()
+
+        return defaults.objectForKey(keyName) as? Memento ?? Memento()
+    }
+}
+/*:
+ ### Usage
+*/
+var gameState = GameState()
+gameState.restoreFromMemento(CheckPoint.restorePreviousState())
+
+gameState.chapter = "Black Mesa Inbound"
+gameState.weapon = "Crowbar"
+CheckPoint.saveState(gameState.toMemento())
+
+gameState.chapter = "Anomalous Materials"
+gameState.weapon = "Glock 17"
+gameState.restoreFromMemento(CheckPoint.restorePreviousState())
+
+gameState.chapter = "Unforeseen Consequences"
+gameState.weapon = "MP5"
+CheckPoint.saveState(gameState.toMemento(), keyName: "gameState2")
+
+gameState.chapter = "Office Complex"
+gameState.weapon = "Crossbow"
+CheckPoint.saveState(gameState.toMemento())
+
+gameState.restoreFromMemento(CheckPoint.restorePreviousState(keyName: "gameState2"))
+
+/*:
+👓 Observer
+-----------
+
+The observer pattern is used to allow an object to publish changes to its state.
+Other objects subscribe to be immediately notified of any changes.
+
+### Example
+*/
+protocol PropertyObserver : class {
+    func willChangePropertyName(propertyName:String, newPropertyValue:AnyObject?)
+    func didChangePropertyName(propertyName:String, oldPropertyValue:AnyObject?)
+}
+
+class TestChambers {
+
+    weak var observer:PropertyObserver?
+
+    var testChamberNumber: Int = 0 {
+        willSet(newValue) {
+            observer?.willChangePropertyName("testChamberNumber", newPropertyValue:newValue)
+        }
+        didSet {
+            observer?.didChangePropertyName("testChamberNumber", oldPropertyValue:oldValue)
+        }
+    }
+}
+
+class Observer : PropertyObserver {
+    func willChangePropertyName(propertyName: String, newPropertyValue: AnyObject?) {
+        if newPropertyValue as? Int == 1 {
+            print("Okay. Look. We both said a lot of things that you're going to regret.")
+        }
+    }
+
+    func didChangePropertyName(propertyName: String, oldPropertyValue: AnyObject?) {
+        if oldPropertyValue as? Int == 0 {
+            print("Sorry about the mess. I've really let the place go since you killed me.")
+        }
+    }
+}
+/*:
+### Usage
+*/
+var observerInstance = Observer()
+var testChambers = TestChambers()
+testChambers.observer = observerInstance
+testChambers.testChamberNumber++
+/*:
+🐉 State
+---------
+
+The state pattern is used to alter the behaviour of an object as its internal state changes.
+The pattern allows the class for an object to apparently change at run-time.
+
+### Example
+*/
+class Context {
+	private var state: State = UnauthorizedState()
+
+    var isAuthorized: Bool {
+        get { return state.isAuthorized(self) }
+    }
+
+    var userId: String? {
+        get { return state.userId(self) }
+    }
+
+	func changeStateToAuthorized(userId userId: String) {
+		state = AuthorizedState(userId: userId)
+	}
+
+	func changeStateToUnauthorized() {
+		state = UnauthorizedState()
+	}
+
+}
+
+protocol State {
+	func isAuthorized(context: Context) -> Bool
+	func userId(context: Context) -> String?
+}
+
+class UnauthorizedState: State {
+	func isAuthorized(context: Context) -> Bool { return false }
+
+	func userId(context: Context) -> String? { return nil }
+}
+
+class AuthorizedState: State {
+	let userId: String
+
+	init(userId: String) { self.userId = userId }
+
+	func isAuthorized(context: Context) -> Bool { return true }
+
+	func userId(context: Context) -> String? { return userId }
+}
+/*:
+### Usage
+*/
+let context = Context()
+(context.isAuthorized, context.userId)
+context.changeStateToAuthorized(userId: "admin")
+(context.isAuthorized, context.userId) // now logged in as "admin"
+context.changeStateToUnauthorized()
+(context.isAuthorized, context.userId)
+/*:
+💡 Strategy
+-----------
+
+The strategy pattern is used to create an interchangeable family of algorithms from which the required process is chosen at run-time.
+
+### Example
+*/
+protocol PrintStrategy {
+    func printString(string: String) -> String
+}
+
+class Printer {
+
+    let strategy: PrintStrategy
+
+    func printString(string: String) -> String {
+        return self.strategy.printString(string)
+    }
+
+    init(strategy: PrintStrategy) {
+        self.strategy = strategy
+    }
+}
+
+class UpperCaseStrategy : PrintStrategy {
+    func printString(string:String) -> String {
+        return string.uppercaseString
+    }
+}
+
+class LowerCaseStrategy : PrintStrategy {
+    func printString(string:String) -> String {
+        return string.lowercaseString
+    }
+}
+/*:
+### Usage
+*/
+var lower = Printer(strategy:LowerCaseStrategy())
+lower.printString("O tempora, o mores!")
+
+var upper = Printer(strategy:UpperCaseStrategy())
+upper.printString("O tempora, o mores!")
+
+/*:
+🏃 Visitor
+----------
+
+The visitor pattern is used to separate a relatively complex set of structured data classes from the functionality that may be performed upon the data that they hold.
+
+### Example
+*/
+protocol PlanetVisitor {
+	func visit(planet: PlanetAlderaan)
+	func visit(planet: PlanetCoruscant)
+	func visit(planet: PlanetTatooine)
+}
+
+protocol Planet {
+	func accept(visitor: PlanetVisitor)
+}
+
+class PlanetAlderaan: Planet {
+	func accept(visitor: PlanetVisitor) { visitor.visit(self) }
+}
+class PlanetCoruscant: Planet {
+	func accept(visitor: PlanetVisitor) { visitor.visit(self) }
+}
+class PlanetTatooine: Planet {
+	func accept(visitor: PlanetVisitor) { visitor.visit(self) }
+}
+
+class NameVisitor: PlanetVisitor {
+	var name = ""
+
+	func visit(planet: PlanetAlderaan)  { name = "Alderaan" }
+	func visit(planet: PlanetCoruscant) { name = "Coruscant" }
+	func visit(planet: PlanetTatooine)  { name = "Tatooine" }
+}
+/*:
+### Usage
+*/
+let planets: [Planet] = [PlanetAlderaan(), PlanetCoruscant(), PlanetTatooine()]
+
+let names = planets.map { (planet: Planet) -> String in
+	let visitor = NameVisitor()
+	planet.accept(visitor)
+	return visitor.name
+}
+
+names
