+++ Design-Patterns.playground/Pages/Structural.xcplaygroundpage/Contents.swift
@@ -1,0 +1,348 @@
+//: [Behavioral](Behavioral) |
+//: [Creational](Creational) |
+//: Structural
+/*:
+Structural
+==========
+
+>In software engineering, structural design patterns are design patterns that ease the design by identifying a simple way to realize relationships between entities.
+>
+>**Source:** [wikipedia.org](http://en.wikipedia.org/wiki/Structural_pattern)
+*/
+import Swift
+import Foundation/*:
+🔌 Adapter
+----------
+
+The adapter pattern is used to provide a link between two otherwise incompatible types by wrapping the "adaptee" with a class that supports the interface required by the client.
+
+### Example
+*/
+protocol OlderDeathStarSuperLaserAiming {
+    var angleV: NSNumber {get}
+    var angleH: NSNumber {get}
+}
+/*:
+**Adaptee**
+*/
+struct DeathStarSuperlaserTarget {
+    let angleHorizontal: Double
+    let angleVertical: Double
+
+    init(angleHorizontal:Double, angleVertical:Double) {
+        self.angleHorizontal = angleHorizontal
+        self.angleVertical = angleVertical
+    }
+}
+/*:
+**Adapter**
+*/
+struct OldDeathStarSuperlaserTarget : OlderDeathStarSuperLaserAiming {
+    private let target : DeathStarSuperlaserTarget
+
+    var angleV:NSNumber {
+        return NSNumber(double: target.angleVertical)
+    }
+
+    var angleH:NSNumber {
+        return NSNumber(double: target.angleHorizontal)
+    }
+
+    init(_ target:DeathStarSuperlaserTarget) {
+        self.target = target
+    }
+}
+/*:
+### Usage
+*/
+let target = DeathStarSuperlaserTarget(angleHorizontal: 14.0, angleVertical: 12.0)
+let oldFormat = OldDeathStarSuperlaserTarget(target)
+
+oldFormat.angleH
+oldFormat.angleV
+/*:
+🌉  Bridge
+----------
+
+The bridge pattern is used to separate the abstract elements of a class from the implementation details, providing the means to replace the implementation details without modifying the abstraction.
+
+### Example
+*/
+protocol Switch {
+    var appliance: Appliance {get set}
+    func turnOn()
+}
+
+protocol Appliance {
+    func run()
+}
+
+class RemoteControl: Switch {
+    var appliance: Appliance
+
+    func turnOn() {
+        self.appliance.run()
+    }
+
+    init(appliance: Appliance) {
+        self.appliance = appliance
+    }
+}
+
+class TV: Appliance {
+    func run() {
+        print("tv turned on");
+    }
+}
+
+class VacuumCleaner: Appliance {
+    func run() {
+        print("vacuum cleaner turned on")
+    }
+}
+/*:
+### Usage
+*/
+var tvRemoteControl = RemoteControl(appliance: TV())
+tvRemoteControl.turnOn()
+
+var fancyVacuumCleanerRemoteControl = RemoteControl(appliance: VacuumCleaner())
+fancyVacuumCleanerRemoteControl.turnOn()
+/*:
+🌿 Composite
+-------------
+
+The composite pattern is used to create hierarchical, recursive tree structures of related objects where any element of the structure may be accessed and utilised in a standard manner.
+
+### Example
+*/
+/*:
+Component
+*/
+protocol Shape {
+    func draw(fillColor: String)
+}
+/*:
+Leafs
+*/
+class Square : Shape {
+    func draw(fillColor: String) {
+        print("Drawing a Square with color \(fillColor)")
+    }
+}
+
+class Circle : Shape {
+    func draw(fillColor: String) {
+        print("Drawing a circle with color \(fillColor)")
+    }
+}
+
+/*:
+Composite
+*/
+class Whiteboard : Shape {
+    lazy var shapes = [Shape]()
+
+    init(_ shapes:Shape...) {
+        self.shapes = shapes
+    }
+
+    func draw(fillColor:String) {
+        for shape in self.shapes {
+            shape.draw(fillColor)
+        }
+    }
+}
+/*:
+### Usage:
+*/
+var whiteboard = Whiteboard(Circle(), Square())
+whiteboard.draw("Red")
+/*:
+🍧 Decorator
+------------
+
+The decorator pattern is used to extend or alter the functionality of objects at run- time by wrapping them in an object of a decorator class.
+This provides a flexible alternative to using inheritance to modify behaviour.
+
+### Example
+*/
+protocol Coffee {
+    func getCost() -> Double
+    func getIngredients() -> String
+}
+
+class SimpleCoffee: Coffee {
+    func getCost() -> Double {
+        return 1.0
+    }
+
+    func getIngredients() -> String {
+        return "Coffee"
+    }
+}
+
+class CoffeeDecorator: Coffee {
+    private let decoratedCoffee: Coffee
+    private let ingredientSeparator: String = ", "
+
+    required init(decoratedCoffee: Coffee) {
+        self.decoratedCoffee = decoratedCoffee
+    }
+
+    func getCost() -> Double {
+        return decoratedCoffee.getCost()
+    }
+
+    func getIngredients() -> String {
+        return decoratedCoffee.getIngredients()
+    }
+}
+
+class Milk: CoffeeDecorator {
+    required init(decoratedCoffee: Coffee) {
+        super.init(decoratedCoffee: decoratedCoffee)
+    }
+
+    override func getCost() -> Double {
+        return super.getCost() + 0.5
+    }
+
+    override func getIngredients() -> String {
+        return super.getIngredients() + ingredientSeparator + "Milk"
+    }
+}
+
+class WhipCoffee: CoffeeDecorator {
+    required init(decoratedCoffee: Coffee) {
+        super.init(decoratedCoffee: decoratedCoffee)
+    }
+
+    override func getCost() -> Double {
+        return super.getCost() + 0.7
+    }
+
+    override func getIngredients() -> String {
+        return super.getIngredients() + ingredientSeparator + "Whip"
+    }
+}
+/*:
+### Usage:
+*/
+var someCoffee: Coffee = SimpleCoffee()
+print("Cost : \(someCoffee.getCost()); Ingredients: \(someCoffee.getIngredients())")
+someCoffee = Milk(decoratedCoffee: someCoffee)
+print("Cost : \(someCoffee.getCost()); Ingredients: \(someCoffee.getIngredients())")
+someCoffee = WhipCoffee(decoratedCoffee: someCoffee)
+print("Cost : \(someCoffee.getCost()); Ingredients: \(someCoffee.getIngredients())")
+/*:
+🎁 Façade
+---------
+
+The facade pattern is used to define a simplified interface to a more complex subsystem.
+
+### Example
+*/
+class Eternal {
+
+    class func setObject(value: AnyObject!, forKey defaultName: String!) {
+        let defaults:NSUserDefaults = NSUserDefaults.standardUserDefaults()
+        defaults.setObject(value, forKey:defaultName)
+        defaults.synchronize()
+    }
+
+    class func objectForKey(defaultName: String!) -> AnyObject! {
+        let defaults:NSUserDefaults = NSUserDefaults.standardUserDefaults()
+
+        return defaults.objectForKey(defaultName)
+    }
+
+}
+/*:
+### Usage
+*/
+Eternal.setObject("Disconnect me. I’d rather be nothing", forKey:"Bishop")
+Eternal.objectForKey("Bishop")
+/*:
+☔ Protection Proxy
+------------------
+
+The proxy pattern is used to provide a surrogate or placeholder object, which references an underlying object.
+Protection proxy is restricting access.
+
+### Example
+*/
+protocol DoorOperator {
+    func openDoors(doors: String) -> String
+}
+
+class HAL9000 : DoorOperator {
+    func openDoors(doors: String) -> String {
+        return ("HAL9000: Affirmative, Dave. I read you. Opened \(doors).")
+    }
+}
+
+class CurrentComputer : DoorOperator {
+    private var computer: HAL9000!
+
+    func authenticateWithPassword(pass: String) -> Bool {
+
+        if pass != "pass" {
+            return false
+        }
+
+        computer = HAL9000()
+
+        return true
+    }
+
+    func openDoors(doors: String) -> String {
+
+        if computer == nil {
+            return "Access Denied. I'm afraid I can't do that."
+        }
+
+        return computer.openDoors(doors)
+    }
+}
+/*:
+### Usage
+*/
+let computer = CurrentComputer()
+let doors = "Pod Bay Doors"
+
+computer.openDoors(doors)
+
+computer.authenticateWithPassword("pass")
+computer.openDoors(doors)
+/*:
+🍬 Virtual Proxy
+----------------
+
+The proxy pattern is used to provide a surrogate or placeholder object, which references an underlying object.
+Virtual proxy is used for loading object on demand.
+
+### Example
+*/
+protocol HEVSuitMedicalAid {
+    func administerMorphine() -> String
+}
+
+class HEVSuit : HEVSuitMedicalAid {
+    func administerMorphine() -> String {
+        return "Morphine aministered."
+    }
+}
+
+class HEVSuitHumanInterface : HEVSuitMedicalAid {
+    lazy private var physicalSuit: HEVSuit = HEVSuit()
+
+    func administerMorphine() -> String {
+        return physicalSuit.administerMorphine()
+    }
+}
+/*:
+### Usage
+*/
+let humanInterface = HEVSuitHumanInterface()
+humanInterface.administerMorphine()
