--- Design-Patterns.playground/contents.swift
@@ -1,1217 +1,0 @@
-
-import Swift
-import Foundation
-
-/*:
-
-Design Patterns implemented in Swift 1.2
-========================================
-A short cheat-sheet with Xcode 6.3beta Playground ([Design-Patterns.playground.zip](https://raw.githubusercontent.com/ochococo/Design-Patterns-In-Swift/master/Design-Patterns.playground.zip)).
-
-👷 Project maintained by: [@nsmeme](http://twitter.com/nsmeme) (Oktawian Chojnacki)
-
-## Table of Contents
-
-* [Behavioral](#behavioral)
-* [Creational](#creational)
-* [Structural](#structural)
-
-*//*:
-Behavioral
-==========
-
->In software engineering, behavioral design patterns are design patterns that identify common communication patterns between objects and realize these patterns. By doing so, these patterns increase flexibility in carrying out this communication.
->
->**Source:** [wikipedia.org](http://en.wikipedia.org/wiki/Behavioral_pattern)
-*//*:
-🐝 Chain Of Responsibility
---------------------------
-
-The chain of responsibility pattern is used to process varied requests, each of which may be dealt with by a different handler.
-
-### Example:
-*/
-class MoneyPile {
-    let value: Int
-    var quantity: Int
-    var nextPile: MoneyPile?
-    
-    init(value: Int, quantity: Int, nextPile: MoneyPile?) {
-        self.value = value
-        self.quantity = quantity
-        self.nextPile = nextPile
-    }
-    
-    func canWithdraw(var v: Int) -> Bool {
-
-        func canTakeSomeBill(want: Int) -> Bool {
-            return (want / self.value) > 0
-        }
-        
-        var q = self.quantity
-
-        while canTakeSomeBill(v) {
-
-            if q == 0 {
-                break
-            }
-
-            v -= self.value
-            q -= 1
-        }
-
-        if v == 0 {
-            return true
-        } else if let next = self.nextPile {
-            return next.canWithdraw(v)
-        }
-
-        return false
-    }
-}
-
-class ATM {
-    private var hundred: MoneyPile
-    private var fifty: MoneyPile
-    private var twenty: MoneyPile
-    private var ten: MoneyPile
-    
-    private var startPile: MoneyPile {
-        return self.hundred
-    }
-    
-    init(hundred: MoneyPile, 
-           fifty: MoneyPile, 
-          twenty: MoneyPile, 
-             ten: MoneyPile) {
-
-        self.hundred = hundred
-        self.fifty = fifty
-        self.twenty = twenty
-        self.ten = ten
-    }
-    
-    func canWithdraw(value: Int) -> String {
-        return "Can withdraw: \(self.startPile.canWithdraw(value))"
-    }
-}
-/*:
-### Usage
-*/
-// Create piles of money and link them together 10 < 20 < 50 < 100.**
-let ten = MoneyPile(value: 10, quantity: 6, nextPile: nil)
-let twenty = MoneyPile(value: 20, quantity: 2, nextPile: ten)
-let fifty = MoneyPile(value: 50, quantity: 2, nextPile: twenty)
-let hundred = MoneyPile(value: 100, quantity: 1, nextPile: fifty)
-
-// Build ATM.
-var atm = ATM(hundred: hundred, fifty: fifty, twenty: twenty, ten: ten)
-atm.canWithdraw(310) // Cannot because ATM has only 300
-atm.canWithdraw(100) // Can withdraw - 1x100
-atm.canWithdraw(165) // Cannot withdraw because ATM doesn't has bill with value of 5
-atm.canWithdraw(30)  // Can withdraw - 1x20, 2x10
-/*:
-👫 Command
-----------
-
-The command pattern is used to express a request, including the call to be made and all of its required parameters, in a command object. The command may then be executed immediately or held for later use.
-
-### Example:
-*/
-protocol DoorCommand {
-    func execute() -> String
-}
-
-class OpenCommand : DoorCommand {
-    let doors:String
-
-    required init(doors: String) {
-        self.doors = doors
-    }
-    
-    func execute() -> String {
-        return "Opened \(doors)"
-    }
-}
-
-class CloseCommand : DoorCommand {
-    let doors:String
-
-    required init(doors: String) {
-        self.doors = doors
-    }
-    
-    func execute() -> String {
-        return "Closed \(doors)"
-    }
-}
-
-class HAL9000DoorsOperations {
-    let openCommand: DoorCommand
-    let closeCommand: DoorCommand
-    
-    init(doors: String) {
-        self.openCommand = OpenCommand(doors:doors)
-        self.closeCommand = CloseCommand(doors:doors)
-    }
-    
-    func close() -> String {
-        return closeCommand.execute()
-    }
-    
-    func open() -> String {
-        return openCommand.execute()
-    }
-}
-/*:
-### Usage:
-*/
-let podBayDoors = "Pod Bay Doors"
-let doorModule = HAL9000DoorsOperations(doors:podBayDoors)
-
-doorModule.open()
-doorModule.close()
-/*:
-🎶 Interpreter
---------------
-
-The interpreter pattern is used to evaluate sentences in a language.
-
-### Example
-*/
-
-protocol IntegerExp {
-    func evaluate(context: IntegerContext) -> Int
-    func replace(character: Character, integerExp: IntegerExp) -> IntegerExp
-    func copy() -> IntegerExp
-}
-
-class IntegerContext {
-    private var data: [Character:Int] = [:]
-    
-    func lookup(name: Character) -> Int {
-        return self.data[name]!
-    }
-    
-    func assign(integerVarExp: IntegerVarExp, value: Int) {
-        self.data[integerVarExp.name] = value
-    }
-}
-
-class IntegerVarExp: IntegerExp {
-    let name: Character
-    
-    init(name: Character) {
-        self.name = name
-    }
-    
-    func evaluate(context: IntegerContext) -> Int {
-        return context.lookup(self.name)
-    }
-    
-    func replace(name: Character, integerExp: IntegerExp) -> IntegerExp {
-        if name == self.name {
-            return integerExp.copy()
-        } else {
-            return IntegerVarExp(name: self.name)
-        }
-    }
-    
-    func copy() -> IntegerExp {
-        return IntegerVarExp(name: self.name)
-    }
-}
-
-class AddExp: IntegerExp {
-    private var operand1: IntegerExp
-    private var operand2: IntegerExp
-    
-    init(op1: IntegerExp, op2: IntegerExp) {
-        self.operand1 = op1
-        self.operand2 = op2
-    }
-    
-    func evaluate(context: IntegerContext) -> Int {
-        return self.operand1.evaluate(context) + self.operand2.evaluate(context)
-    }
-    
-    func replace(character: Character, integerExp: IntegerExp) -> IntegerExp {
-        return AddExp(op1: operand1.replace(character, integerExp: integerExp),
-            op2: operand2.replace(character, integerExp: integerExp))
-    }
-    
-    func copy() -> IntegerExp {
-        return AddExp(op1: self.operand1, op2: self.operand2)
-    }
-}
-/*:
-### Usage
-*/
-var expression: IntegerExp?
-var intContext = IntegerContext()
-
-var a = IntegerVarExp(name: "A")
-var b = IntegerVarExp(name: "B")
-var c = IntegerVarExp(name: "C")
-
-expression = AddExp(op1: a, op2: AddExp(op1: b, op2: c)) // a + (b + c)
-
-intContext.assign(a, value: 2)
-intContext.assign(b, value: 1)
-intContext.assign(c, value: 3)
-
-var result = expression?.evaluate(intContext)
-/*:
-🍫 Iterator
------------
-
-The iterator pattern is used to provide a standard interface for traversing a collection of items in an aggregate object without the need to understand its underlying structure.
-
-### Example:
-*/
-struct NovellasCollection<T> {
-    let novellas: [T]
-}
-
-extension NovellasCollection: SequenceType {
-    typealias Generator = GeneratorOf<T>
-    
-    func generate() -> GeneratorOf<T> {
-        var i = 0
-        return GeneratorOf { return i >= self.novellas.count ? nil : self.novellas[i++] }
-    }
-}
-/*:
-### Usage
-*/
-let greatNovellas = NovellasCollection(novellas:["Mist"])
-
-for novella in greatNovellas {
-    println("I've read: \(novella)")
-}
-/*:
-💐 Mediator
------------
-
-The mediator pattern is used to reduce coupling between classes that communicate with each other. Instead of classes communicating directly, and thus requiring knowledge of their implementation, the classes send messages via a mediator object.
-
-### Example
-*/
-
-class Colleague {
-    let mediator: Mediator
-    
-    init(mediator: Mediator) {
-        self.mediator = mediator
-    }
-    
-    func send(message: String) {
-        mediator.send(message, colleague: self)
-    }
-    
-    func receive(message: String) {
-        assert(false, "Method should be overriden")
-    }
-}
-
-protocol Mediator {
-    func send(message: String, colleague: Colleague)
-}
-
-class MessageMediator: Mediator {
-    private var colleagues: [Colleague] = []
-    
-    func addColleague(colleague: Colleague) {
-        colleagues.append(colleague)
-    }
-    
-    func send(message: String, colleague: Colleague) {
-        for c in colleagues {
-            if c !== colleague { //for simplicity we compare object references
-                colleague.receive(message)
-            }
-        }
-    }
-}
-
-class ConcreteColleague: Colleague {
-    override func receive(message: String) {
-        println("Colleague received: \(message)")
-    }
-}
-
-/*:
-### Usage
-*/
-
-let messagesMediator = MessageMediator()
-let user0 = ConcreteColleague(mediator: messagesMediator)
-let user1 = ConcreteColleague(mediator: messagesMediator)
-messagesMediator.addColleague(user0)
-messagesMediator.addColleague(user1)
-
-user0.send("Hello") // user1 receives message
-/*:
-💾 Memento
-----------
-
-The memento pattern is used to capture the current state of an object and store it in such a manner that it can be restored at a later time without breaking the rules of encapsulation.
-
-### Example
-*/
-typealias Memento = Dictionary<NSObject, AnyObject>
-
-let DPMementoKeyChapter = "com.valve.halflife.chapter"
-let DPMementoKeyWeapon = "com.valve.halflife.weapon"
-let DPMementoGameState = "com.valve.halflife.state"
-/*:
-Originator
-*/
-class GameState {
-    var chapter: String = ""
-    var weapon: String = ""
-
-    func toMemento() -> Memento {
-        return [ DPMementoKeyChapter:chapter, DPMementoKeyWeapon:weapon ]
-    }
-
-    func restoreFromMemento(memento: Memento) {
-        chapter = memento[DPMementoKeyChapter] as? String ?? "n/a"
-        weapon = memento[DPMementoKeyWeapon] as? String ?? "n/a"
-    }
-}
-/*:
-Caretaker
-*/
-class CheckPoint {
-    class func saveState(memento: Memento, keyName: String = DPMementoGameState) {
-        let defaults = NSUserDefaults.standardUserDefaults()
-        defaults.setObject(memento, forKey: keyName)
-        defaults.synchronize()
-    }
-
-    class func restorePreviousState(keyName: String = DPMementoGameState) -> Memento {
-        let defaults = NSUserDefaults.standardUserDefaults()
-
-        return defaults.objectForKey(keyName) as? Memento ?? Memento()
-    }
-}
-/*:
- ### Usage
-*/
-var gameState = GameState()
-gameState.restoreFromMemento(CheckPoint.restorePreviousState())
-
-gameState.chapter = "Black Mesa Inbound"
-gameState.weapon = "Crowbar"
-CheckPoint.saveState(gameState.toMemento())
-
-gameState.chapter = "Anomalous Materials"
-gameState.weapon = "Glock 17"
-gameState.restoreFromMemento(CheckPoint.restorePreviousState())
-
-gameState.chapter = "Unforeseen Consequences"
-gameState.weapon = "MP5"
-CheckPoint.saveState(gameState.toMemento(), keyName: "gameState2")
-
-gameState.chapter = "Office Complex"
-gameState.weapon = "Crossbow"
-CheckPoint.saveState(gameState.toMemento())
-
-gameState.restoreFromMemento(CheckPoint.restorePreviousState(keyName: "gameState2"))
-
-/*:
-👓 Observer
------------
-
-The observer pattern is used to allow an object to publish changes to its state. 
-Other objects subscribe to be immediately notified of any changes.
-
-### Example
-*/
-protocol PropertyObserver : class {
-    func willChangePropertyName(propertyName:String, newPropertyValue:AnyObject?)
-    func didChangePropertyName(propertyName:String, oldPropertyValue:AnyObject?)
-}
-
-class TestChambers {
-
-    weak var observer:PropertyObserver?
-
-    var testChamberNumber: Int = 0 {
-        willSet(newValue) {
-            observer?.willChangePropertyName("testChamberNumber", newPropertyValue:newValue)
-        }
-        didSet {
-            observer?.didChangePropertyName("testChamberNumber", oldPropertyValue:oldValue)
-        }
-    }
-}
-
-class Observer : PropertyObserver {
-    func willChangePropertyName(propertyName: String, newPropertyValue: AnyObject?) {
-        if newPropertyValue as? Int == 1 {
-            println("Okay. Look. We both said a lot of things that you're going to regret.")
-        }
-    }
-
-    func didChangePropertyName(propertyName: String, oldPropertyValue: AnyObject?) {
-        if oldPropertyValue as? Int == 0 {
-            println("Sorry about the mess. I've really let the place go since you killed me.")
-        }
-    }
-}
-/*:
-### Usage
-*/
-var observerInstance = Observer()
-var testChambers = TestChambers()
-testChambers.observer = observerInstance
-testChambers.testChamberNumber++
-/*:
-## 🐉 State
-
-The state pattern is used to alter the behaviour of an object as its internal state changes. 
-The pattern allows the class for an object to apparently change at run-time.
-
-### Example
-*/
-class Context {
-	private var state: State = UnauthorizedState()
-
-    var isAuthorized: Bool {
-        get { return state.isAuthorized(self) }
-    }
-
-    var userId: String? {
-        get { return state.userId(self) }
-    }
-
-	func changeStateToAuthorized(#userId: String) {
-		state = AuthorizedState(userId: userId)
-	}
-
-	func changeStateToUnauthorized() {
-		state = UnauthorizedState()
-	}
-    
-}
-
-protocol State {
-	func isAuthorized(context: Context) -> Bool
-	func userId(context: Context) -> String?
-}
-
-class UnauthorizedState: State {
-	func isAuthorized(context: Context) -> Bool { return false }
-
-	func userId(context: Context) -> String? { return nil }
-}
-
-class AuthorizedState: State {
-	let userId: String
-
-	init(userId: String) { self.userId = userId }
-
-	func isAuthorized(context: Context) -> Bool { return true }
-
-	func userId(context: Context) -> String? { return userId }
-}
-/*:
-### Usage
-*/
-let context = Context()
-(context.isAuthorized, context.userId)
-context.changeStateToAuthorized(userId: "admin")
-(context.isAuthorized, context.userId) // now logged in as "admin"
-context.changeStateToUnauthorized()
-(context.isAuthorized, context.userId)
-/*:
-💡 Strategy
------------
-
-The strategy pattern is used to create an interchangeable family of algorithms from which the required process is chosen at run-time.
-
-### Example
-*/
-protocol PrintStrategy {
-    func printString(string: String) -> String
-}
-
-class Printer {
-
-    let strategy: PrintStrategy
-    
-    func printString(string: String) -> String {
-        return self.strategy.printString(string)
-    }
-    
-    init(strategy: PrintStrategy) {
-        self.strategy = strategy
-    }
-}
-
-class UpperCaseStrategy : PrintStrategy {
-    func printString(string:String) -> String {
-        return string.uppercaseString
-    }
-}
-
-class LowerCaseStrategy : PrintStrategy {
-    func printString(string:String) -> String {
-        return string.lowercaseString
-    }
-}
-/*:
-### Usage
-*/
-var lower = Printer(strategy:LowerCaseStrategy())
-lower.printString("O tempora, o mores!")
-
-var upper = Printer(strategy:UpperCaseStrategy())
-upper.printString("O tempora, o mores!")
-
-/*:
-🏃 Visitor
-----------
-
-The visitor pattern is used to separate a relatively complex set of structured data classes from the functionality that may be performed upon the data that they hold.
-
-### Example
-*/
-protocol PlanetVisitor {
-	func visit(planet: PlanetAlderaan)
-	func visit(planet: PlanetCoruscant)
-	func visit(planet: PlanetTatooine)
-}
-
-protocol Planet {
-	func accept(visitor: PlanetVisitor)
-}
-
-class PlanetAlderaan: Planet {
-	func accept(visitor: PlanetVisitor) { visitor.visit(self) }
-}
-class PlanetCoruscant: Planet {
-	func accept(visitor: PlanetVisitor) { visitor.visit(self) }
-}
-class PlanetTatooine: Planet {
-	func accept(visitor: PlanetVisitor) { visitor.visit(self) }
-}
-
-class NameVisitor: PlanetVisitor {
-	var name = ""
-
-	func visit(planet: PlanetAlderaan)  { name = "Alderaan" }
-	func visit(planet: PlanetCoruscant) { name = "Coruscant" }
-	func visit(planet: PlanetTatooine)  { name = "Tatooine" }
-}
-/*:
-### Usage
-*/
-let planets: [Planet] = [PlanetAlderaan(), PlanetCoruscant(), PlanetTatooine()]
-
-let names = planets.map { (planet: Planet) -> String in
-	let visitor = NameVisitor()
-	planet.accept(visitor)
-	return visitor.name
-}
-
-names
-/*:
-Creational
-==========
-
-> In software engineering, creational design patterns are design patterns that deal with object creation mechanisms, trying to create objects in a manner suitable to the situation. The basic form of object creation could result in design problems or added complexity to the design. Creational design patterns solve this problem by somehow controlling this object creation.
->
->**Source:** [wikipedia.org](http://en.wikipedia.org/wiki/Creational_pattern)
-*//*:
-🌰 Abstract Factory
--------------------
-
-The abstract factory pattern is used to provide a client with a set of related or dependant objects. 
-The "family" of objects created by the factory are determined at run-time.
-
-### Example
-*/
-/*: 
-Protocols
-*/
-protocol Decimal {
-    func stringValue() -> String
-}
-
-protocol NumberFactoryProtocol {
-    func numberFromString(string : String) -> Decimal
-}
-
-// Number implementations
-
-struct NextStepNumber : Decimal {
-    private var nextStepNumber : NSNumber
-
-    func stringValue() -> String { return nextStepNumber.stringValue }
-}
-
-struct SwiftNumber : Decimal {
-    private var swiftInt : Int
-
-    func stringValue() -> String { return "\(swiftInt)" }
-}
-/*:
-Factories
-*/
-class NextStepNumberFactory : NumberFactoryProtocol {
-    func numberFromString(string : String) -> Decimal {
-        return NextStepNumber(nextStepNumber:NSNumber(longLong:(string as NSString).longLongValue))
-    }
-}
-
-class SwiftNumberFactory : NumberFactoryProtocol {
-    func numberFromString(string : String) -> Decimal {
-        return SwiftNumber(swiftInt:(string as NSString).integerValue)
-    }
-}
-/*:
-Abstract factory
-*/
-enum NumberType {
-    case NextStep, Swift
-}
-
-class NumberAbstractFactory {
-    class func numberFactoryType(type : NumberType) -> NumberFactoryProtocol {
-        
-        switch type {
-            case .NextStep:
-                    return NextStepNumberFactory()
-            case .Swift:
-                    return SwiftNumberFactory()
-        }
-    }
-}
-/*:
-### Usage
-*/
-let factoryOne = NumberAbstractFactory.numberFactoryType(.NextStep)
-let numberOne = factoryOne.numberFromString("1")
-numberOne.stringValue()
-
-let factoryTwo = NumberAbstractFactory.numberFactoryType(.Swift)
-let numberTwo = factoryTwo.numberFromString("2")
-numberTwo.stringValue()
-/*:
-👷 Builder
-----------
-
-The builder pattern is used to create complex objects with constituent parts that must be created in the same order or using a specific algorithm. 
-An external class controls the construction algorithm.
-
-### Example
-*/
-class DeathStarBuilder {
-
-    var x: Double?
-    var y: Double?
-    var z: Double?
-
-    typealias BuilderClosure = (DeathStarBuilder) -> ()
-
-    init(buildClosure: BuilderClosure) {
-        buildClosure(self)
-    }
-}
-
-struct DeathStar {
-
-    let x: Double
-    let y: Double
-    let z: Double
-
-    init?(builder: DeathStarBuilder) {
-
-        if let x = builder.x, y = builder.y, z = builder.z {
-            self.x = x
-            self.y = y
-            self.z = z
-        } else {
-            return nil
-        }
-    }
-}
-/*:
-### Usage
-*/
-let empire = DeathStarBuilder { builder in
-    builder.x = 0.1
-    builder.y = 0.2
-    builder.z = 0.3
-}
-
-let deathStar = DeathStar(builder:empire)
-/*:
-🏭 Factory Method
------------------
-
-The factory pattern is used to replace class constructors, abstracting the process of object generation so that the type of the object instantiated can be determined at run-time.
-
-### Example
-*/
-protocol Currency {
-    func symbol() -> String
-    func code() -> String
-}
-
-class Euro : Currency {
-    func symbol() -> String {
-        return "€"
-    }
-    
-    func code() -> String {
-        return "EUR"
-    }
-}
-
-class UnitedStatesDolar : Currency {
-    func symbol() -> String {
-        return "$"
-    }
-    
-    func code() -> String {
-        return "USD"
-    }
-}
-
-enum Country {
-    case UnitedStates, Spain, France, UK
-}
-
-class CurrencyFactory {
-    class func currencyForCountry(country:Country) -> Currency? {
-
-        switch country {
-            case .Spain, .France :
-                return Euro()
-            case .UnitedStates :
-                return UnitedStatesDolar()
-            default:
-                return nil
-        }
-        
-    }
-}
-/*:
-### Usage
-*/
-let noCurrencyCode = "No Currency Code Available"
-
-CurrencyFactory.currencyForCountry(.Spain)?.code() ?? noCurrencyCode
-CurrencyFactory.currencyForCountry(.UnitedStates)?.code() ?? noCurrencyCode
-CurrencyFactory.currencyForCountry(.France)?.code() ?? noCurrencyCode
-CurrencyFactory.currencyForCountry(.UK)?.code() ?? noCurrencyCode
-/*:
-🃏 Prototype
-------------
-
-The prototype pattern is used to instantiate a new object by copying all of the properties of an existing object, creating an independent clone. 
-This practise is particularly useful when the construction of a new object is inefficient.
-
-### Example
-*/
-class ChungasRevengeDisplay {
-    var name: String?
-    let font: String
-
-    init(font: String) {
-        self.font = font
-    }
-
-    func clone() -> ChungasRevengeDisplay {
-        return ChungasRevengeDisplay(font:self.font)
-    }
-}
-/*:
-### Usage
-*/
-let Prototype = ChungasRevengeDisplay(font:"GotanProject")
-
-let Philippe = Prototype.clone()
-Philippe.name = "Philippe"
-
-let Christoph = Prototype.clone()
-Christoph.name = "Christoph"
-
-let Eduardo = Prototype.clone()
-Eduardo.name = "Eduardo"
-/*:
-💍 Singleton
-------------
-
-The singleton pattern ensures that only one object of a particular class is ever created.
-All further references to objects of the singleton class refer to the same underlying instance.
-There are very few applications, do not overuse this pattern!
-
-### Example:
-*/
-class DeathStarSuperlaser {
-    static let sharedInstance = DeathStarSuperlaser()
-
-    private init() {
-        // Private initialization to ensure just one instance is created.
-    }
-}
-/*:
-### Usage:
-*/
-let laser = DeathStarSuperlaser.sharedInstance
-/*:
-Structural
-==========
-
->In software engineering, structural design patterns are design patterns that ease the design by identifying a simple way to realize relationships between entities.
->
->**Source:** [wikipedia.org](http://en.wikipedia.org/wiki/Structural_pattern)
-*//*:
-🔌 Adapter
-----------
-
-The adapter pattern is used to provide a link between two otherwise incompatible types by wrapping the "adaptee" with a class that supports the interface required by the client.
-
-### Example
-*/
-protocol OlderDeathStarSuperLaserAiming {
-    var angleV: NSNumber {get}
-    var angleH: NSNumber {get}
-}
-/*:
-**Adaptee**
-*/
-struct DeathStarSuperlaserTarget {
-    let angleHorizontal: Double
-    let angleVertical: Double
-
-    init(angleHorizontal:Double, angleVertical:Double) {
-        self.angleHorizontal = angleHorizontal
-        self.angleVertical = angleVertical
-    }
-}
-/*:
-**Adapter**
-*/
-struct OldDeathStarSuperlaserTarget : OlderDeathStarSuperLaserAiming {
-    private let target : DeathStarSuperlaserTarget
-
-    var angleV:NSNumber {
-        return NSNumber(double: target.angleVertical)
-    }
-
-    var angleH:NSNumber {
-        return NSNumber(double: target.angleHorizontal)
-    }
-
-    init(_ target:DeathStarSuperlaserTarget) {
-        self.target = target
-    }
-}
-/*:
-### Usage
-*/
-let target = DeathStarSuperlaserTarget(angleHorizontal: 14.0, angleVertical: 12.0)
-let oldFormat = OldDeathStarSuperlaserTarget(target)
-
-oldFormat.angleH
-oldFormat.angleV
-/*:
-🌉  Bridge
-----------
-
-The bridge pattern is used to separate the abstract elements of a class from the implementation details, providing the means to replace the implementation details without modifying the abstraction.
-
-### Example
-*/
-protocol Switch {
-    var appliance: Appliance {get set}
-    func turnOn()
-}
-
-protocol Appliance {
-    func run()
-}
-
-class RemoteControl: Switch {
-    var appliance: Appliance
-
-    func turnOn() {
-        self.appliance.run()
-    }
-    
-    init(appliance: Appliance) {
-        self.appliance = appliance
-    }
-}
-
-class TV: Appliance {
-    func run() {
-        println("tv turned on");
-    }
-}
-
-class VacuumCleaner: Appliance {
-    func run() {
-        println("vacuum cleaner turned on")
-    }
-}
-/*:
-### Usage
-*/
-var tvRemoteControl = RemoteControl(appliance: TV())
-tvRemoteControl.turnOn()
-
-var fancyVacuumCleanerRemoteControl = RemoteControl(appliance: VacuumCleaner())
-fancyVacuumCleanerRemoteControl.turnOn()
-/*:
-## 🌿 Composite
-
-The composite pattern is used to create hierarchical, recursive tree structures of related objects where any element of the structure may be accessed and utilised in a standard manner.
-
-### Example
-*/
-/*:
-Component
-*/
-protocol Shape {
-    func draw(fillColor: String)
-}
-/*: 
-Leafs
-*/ 
-class Square : Shape {
-    func draw(fillColor: String) {
-        print("Drawing a Square with color \(fillColor)")
-    }
-}
-
-class Circle : Shape {
-    func draw(fillColor: String) {
-        print("Drawing a circle with color \(fillColor)")
-    }
-}
-
-/*:
-Composite
-*/
-class Whiteboard : Shape {
-    lazy var shapes = [Shape]()
-    
-    init(_ shapes:Shape...) {
-        self.shapes = shapes
-    }
-    
-    func draw(fillColor:String) {
-        for shape in self.shapes {
-            shape.draw(fillColor)
-        }
-    }
-}
-/*:
-### Usage:
-*/
-var whiteboard = Whiteboard(Circle(), Square())
-whiteboard.draw("Red")
-/*:
-🍧 Decorator
-------------
-
-The decorator pattern is used to extend or alter the functionality of objects at run- time by wrapping them in an object of a decorator class. 
-This provides a flexible alternative to using inheritance to modify behaviour.
-
-### Example
-*/
-protocol Coffee {
-    func getCost() -> Double
-    func getIngredients() -> String
-}
-
-class SimpleCoffee: Coffee {
-    func getCost() -> Double {
-        return 1.0
-    }
-
-    func getIngredients() -> String {
-        return "Coffee"
-    }
-}
-
-class CoffeeDecorator: Coffee {
-    private let decoratedCoffee: Coffee
-    private let ingredientSeparator: String = ", "
-
-    required init(decoratedCoffee: Coffee) {
-        self.decoratedCoffee = decoratedCoffee
-    }
-
-    func getCost() -> Double {
-        return decoratedCoffee.getCost()
-    }
-
-    func getIngredients() -> String {
-        return decoratedCoffee.getIngredients()
-    }
-}
-
-class Milk: CoffeeDecorator {
-    required init(decoratedCoffee: Coffee) {
-        super.init(decoratedCoffee: decoratedCoffee)
-    }
-
-    override func getCost() -> Double {
-        return super.getCost() + 0.5
-    }
-
-    override func getIngredients() -> String {
-        return super.getIngredients() + ingredientSeparator + "Milk"
-    }
-}
-
-class WhipCoffee: CoffeeDecorator {
-    required init(decoratedCoffee: Coffee) {
-        super.init(decoratedCoffee: decoratedCoffee)
-    }
-
-    override func getCost() -> Double {
-        return super.getCost() + 0.7
-    }
-
-    override func getIngredients() -> String {
-        return super.getIngredients() + ingredientSeparator + "Whip"
-    }
-}
-/*:
-### Usage:
-*/
-var someCoffee: Coffee = SimpleCoffee()
-println("Cost : \(someCoffee.getCost()); Ingredients: \(someCoffee.getIngredients())")
-someCoffee = Milk(decoratedCoffee: someCoffee)
-println("Cost : \(someCoffee.getCost()); Ingredients: \(someCoffee.getIngredients())")
-someCoffee = WhipCoffee(decoratedCoffee: someCoffee)
-println("Cost : \(someCoffee.getCost()); Ingredients: \(someCoffee.getIngredients())")
-/*:
-🎁 Façade
----------
-
-The facade pattern is used to define a simplified interface to a more complex subsystem.
-
-### Example
-*/
-class Eternal {
-
-    class func setObject(value: AnyObject!, forKey defaultName: String!) {
-        let defaults:NSUserDefaults = NSUserDefaults.standardUserDefaults()
-        defaults.setObject(value, forKey:defaultName)
-        defaults.synchronize()
-    }
-
-    class func objectForKey(defaultName: String!) -> AnyObject! {
-        let defaults:NSUserDefaults = NSUserDefaults.standardUserDefaults()
-
-        return defaults.objectForKey(defaultName)
-    }
-
-}
-/*:
-### Usage
-*/
-Eternal.setObject("Disconnect me. I’d rather be nothing", forKey:"Bishop")
-Eternal.objectForKey("Bishop")
-/*:
-☔ Protection Proxy
-------------------
-
-The proxy pattern is used to provide a surrogate or placeholder object, which references an underlying object. 
-Protection proxy is restricting access.
-
-### Example
-*/
-protocol DoorOperator {
-    func openDoors(doors: String) -> String
-}
-
-class HAL9000 : DoorOperator {
-    func openDoors(doors: String) -> String {
-        return ("HAL9000: Affirmative, Dave. I read you. Opened \(doors).")
-    }
-}
-
-class CurrentComputer : DoorOperator {
-    private var computer: HAL9000!
-
-    func authenticateWithPassword(pass: String) -> Bool {
-
-        if pass != "pass" {
-            return false
-        }
-
-        computer = HAL9000()
-
-        return true
-    }
-    
-    func openDoors(doors: String) -> String {
-
-        if computer == nil {
-            return "Access Denied. I'm afraid I can't do that."
-        }
-        
-        return computer.openDoors(doors)
-    }
-}
-/*:
-### Usage
-*/
-let computer = CurrentComputer()
-let doors = "Pod Bay Doors"
-
-computer.openDoors(doors)
-
-computer.authenticateWithPassword("pass")
-computer.openDoors(doors)
-/*:
-##🍬 Virtual Proxy
-
-The proxy pattern is used to provide a surrogate or placeholder object, which references an underlying object. 
-Virtual proxy is used for loading object on demand.
-
-### Example
-*/
-protocol HEVSuitMedicalAid {
-    func administerMorphine() -> String
-}
-
-class HEVSuit : HEVSuitMedicalAid {
-    func administerMorphine() -> String {
-        return "Morphine aministered."
-    }
-}
-
-class HEVSuitHumanInterface : HEVSuitMedicalAid {
-    lazy private var physicalSuit: HEVSuit = HEVSuit()
-
-    func administerMorphine() -> String {
-        return physicalSuit.administerMorphine()
-    }
-}
-/*:
-### Usage
-*/
-let humanInterface = HEVSuitHumanInterface()
-humanInterface.administerMorphine()
-/*:
-
-Info
-====
-
-📖 Descriptions from: [Gang of Four Design Patterns Reference Sheet](http://www.blackwasp.co.uk/GangOfFour.aspx)
-
-🚀 How to generate playground (+zip) from source: [GENERATE.md](https://github.com/ochococo/Design-Patterns-In-Swift/blob/master/GENERATE.md)
-
-*/
