+++ Design-Patterns.playground/contents.xcplayground
@@ -1,4 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='5.0' target-platform='ios' display-mode='rendered'>
-    <timeline fileName='timeline.xctimeline'/>
+<playground version='6.0' target-platform='ios' requires-full-environment='true' display-mode='rendered'>
+    <pages>
+        <page name='Patterns'/>
+        <page name='Behavioral'/>
+        <page name='Creational'/>
+        <page name='Structural'/>
+    </pages>
 </playground>
