+++ Design-Patterns.playground/playground.xcworkspace/xcshareddata/Design-Patterns.xccheckout
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
+	<false/>
+	<key>IDESourceControlProjectIdentifier</key>
+	<string>6E6A675F-9997-43AE-BB1C-8E4B69F9AD36</string>
+	<key>IDESourceControlProjectName</key>
+	<string>Design-Patterns</string>
+	<key>IDESourceControlProjectOriginsDictionary</key>
+	<dict>
+		<key>C96876E4909DB6FA99FAEC5AD77808104236AA02</key>
+		<string>https://github.com/ochococo/Design-Patterns-In-Swift.git</string>
+	</dict>
+	<key>IDESourceControlProjectPath</key>
+	<string>Design-Patterns.playground</string>
+	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
+	<dict>
+		<key>C96876E4909DB6FA99FAEC5AD77808104236AA02</key>
+		<string>..</string>
+	</dict>
+	<key>IDESourceControlProjectURL</key>
+	<string>https://github.com/ochococo/Design-Patterns-In-Swift.git</string>
+	<key>IDESourceControlProjectVersion</key>
+	<integer>111</integer>
+	<key>IDESourceControlProjectWCCIdentifier</key>
+	<string>C96876E4909DB6FA99FAEC5AD77808104236AA02</string>
+	<key>IDESourceControlProjectWCConfigurations</key>
+	<array>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>C96876E4909DB6FA99FAEC5AD77808104236AA02</string>
+			<key>IDESourceControlWCCName</key>
+			<string>Design-Patterns-In-Swift</string>
+		</dict>
+	</array>
+</dict>
+</plist>
