+++ source/behavioral/state.swift
@@ -1,5 +1,6 @@
 /*:
-## 🐉 State
+🐉 State
+---------
 
 The state pattern is used to alter the behaviour of an object as its internal state changes. 
 The pattern allows the class for an object to apparently change at run-time.
@@ -17,7 +18,7 @@ class Context {
         get { return state.userId(self) }
     }
 
-	func changeStateToAuthorized(#userId: String) {
+	func changeStateToAuthorized(userId userId: String) {
 		state = AuthorizedState(userId: userId)
 	}
