+++ source/header.swift
@@ -1,12 +1,7 @@
 
-import Swift
-import Foundation
-
-/*:
-
-Design Patterns implemented in Swift 1.2
-========================================
-A short cheat-sheet with Xcode 6.3beta Playground ([Design-Patterns.playground.zip](https://raw.githubusercontent.com/ochococo/Design-Patterns-In-Swift/master/Design-Patterns.playground.zip)).
+Design Patterns implemented in Swift 2
+======================================
+A short cheat-sheet with Xcode 7beta Playground ([Design-Patterns.playground.zip](https://raw.githubusercontent.com/ochococo/Design-Patterns-In-Swift/master/Design-Patterns.playground.zip)).
 
 👷 Project maintained by: [@nsmeme](http://twitter.com/nsmeme) (Oktawian Chojnacki)
