+++ source/structural/composite.swift
@@ -1,5 +1,6 @@
 /*:
-## 🌿 Composite
+🌿 Composite
+-------------
 
 The composite pattern is used to create hierarchical, recursive tree structures of related objects where any element of the structure may be accessed and utilised in a standard manner.
