+++ source/structural/virtual_proxy.swift
@@ -1,5 +1,6 @@
 /*:
-##🍬 Virtual Proxy
+🍬 Virtual Proxy
+----------------
 
 The proxy pattern is used to provide a surrogate or placeholder object, which references an underlying object. 
 Virtual proxy is used for loading object on demand.
